+++ src/otx/algo/classification/backbones/timm.py
@@ -9,41 +9,21 @@ Original papers:
 """
 from __future__ import annotations
 
-from pathlib import Path
 from typing import Literal
 
 import timm
 import torch
 from torch import nn
 
-from otx.algo.utils.mmengine_utils import load_checkpoint_to_model, load_from_http
-
-PRETRAINED_ROOT = "https://github.com/rwightman/pytorch-image-models/releases/download/v0.1-effv2-weights/"
-pretrained_urls = {
-    "efficientnetv2_s_21k": PRETRAINED_ROOT + "tf_efficientnetv2_s_21k-6337ad01.pth",
-    "efficientnetv2_s_1k": PRETRAINED_ROOT + "tf_efficientnetv2_s_21ft1k-d7dafa41.pth",
-}
-
-TIMM_MODEL_NAME_DICT = {
-    "mobilenetv3_large_21k": "mobilenetv3_large_100_miil_in21k",
-    "mobilenetv3_large_1k": "mobilenetv3_large_100_miil",
-    "tresnet": "tresnet_m",
-    "efficientnetv2_s_21k": "tf_efficientnetv2_s_in21k",
-    "efficientnetv2_s_1k": "tf_efficientnetv2_s_in21ft1k",
-    "efficientnetv2_m_21k": "tf_efficientnetv2_m_in21k",
-    "efficientnetv2_m_1k": "tf_efficientnetv2_m_in21ft1k",
-    "efficientnetv2_b0": "tf_efficientnetv2_b0",
-}
-
 TimmModelType = Literal[
-    "mobilenetv3_large_21k",
-    "mobilenetv3_large_1k",
-    "tresnet",
-    "efficientnetv2_s_21k",
-    "efficientnetv2_s_1k",
-    "efficientnetv2_m_21k",
-    "efficientnetv2_m_1k",
-    "efficientnetv2_b0",
+    "mobilenetv3_large_100_miil_in21k",
+    "mobilenetv3_large_100_miil",
+    "tresnet_m",
+    "tf_efficientnetv2_s.in21k",
+    "tf_efficientnetv2_s.in21ft1k",
+    "tf_efficientnetv2_m.in21k",
+    "tf_efficientnetv2_m.in21ft1k",
+    "tf_efficientnetv2_b0",
 ]
 
 
@@ -59,12 +39,15 @@ class TimmBackbone(nn.Module):
     ):
         super().__init__(**kwargs)
         self.backbone = backbone
-        self.pretrained = pretrained
+        self.pretrained: bool | dict = pretrained
         self.is_mobilenet = backbone.startswith("mobilenet")
 
-        self.model = timm.create_model(TIMM_MODEL_NAME_DICT[self.backbone], pretrained=pretrained, num_classes=1000)
-        if self.pretrained:
-            print(f"init weight - {pretrained_urls[self.backbone]}")
+        self.model = timm.create_model(
+            self.backbone,
+            pretrained=pretrained,
+            num_classes=1000,
+        )
+
         self.model.classifier = None  # Detach classifier. Only use 'backbone' part in otx.
         self.num_head_features = self.model.num_features
         self.num_features = self.model.conv_head.in_channels if self.is_mobilenet else self.model.num_features
@@ -97,15 +80,3 @@ class TimmBackbone(nn.Module):
                 param_dict["lr"] = lrs
 
         return parameters
-
-    def init_weights(self, pretrained: str | bool | None = None) -> None:
-        """Initialize weights."""
-        checkpoint = None
-        if isinstance(pretrained, str) and Path(pretrained).exists():
-            checkpoint = torch.load(pretrained, None)
-            print(f"init weight - {pretrained}")
-        elif pretrained is not None:
-            checkpoint = load_from_http(pretrained_urls[self.key])
-            print(f"init weight - {pretrained_urls[self.key]}")
-        if checkpoint is not None:
-            load_checkpoint_to_model(self, checkpoint)