+++ CHANGELOG.md
@@ -68,6 +68,8 @@ All notable changes to this project will be documented in this file.
   (<https://github.com/openvinotoolkit/training_extensions/pull/3942>)
 - Change categories mapping logic
   (<https://github.com/openvinotoolkit/training_extensions/pull/3946>)
+- Fix config converter for tiling
+  (<https://github.com/openvinotoolkit/training_extensions/pull/3973>)
 
 ## \[v2.1.0\]